Added features  to the map
Added so that all markers are shown on the map from the start. Added onClickListener for the markers that toasts distance to the marker from the current location, and also toasts "number of chargers". Added functionality to the location icon which now sends the camera to the current user location