Added functionality to the map
Added all markers on the map at the start of the application. Added functionality to the location button which now sends the camera to the current user location. Added onClickListeners to the markers which toast the distance from the current locaiton aswell as how many chargers are at that station.