VerifyActivity is 90% working
verify activity was not declared in the Manifest file, so I did that, and the Response.code() was not valid, so I fixed that through adding a Response to the function in the ApiInterface.
Debugged and I am not getting any further after I click verify. "end of line something...